Remove old backend and its tests
