+++ CMakeLists.txt
@@ -41,7 +41,6 @@ target_sources(clio PRIVATE
   src/main/impl/Build.cpp
   ## Backend
   src/backend/BackendInterface.cpp
-  src/backend/CassandraBackend.cpp
   src/backend/SimpleCache.cpp
   ## NextGen Backend
   src/backend/cassandra/impl/Future.cpp
@@ -106,7 +105,6 @@ if(BUILD_TESTS)
   set(TEST_TARGET clio_tests)
   add_executable(${TEST_TARGET}
     unittests/Playground.cpp
-    unittests/Backend.cpp
     unittests/Logger.cpp
     unittests/Config.cpp
     unittests/ProfilerTest.cpp
